pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = ""
authors = [
    {name = "Yasmin Venceslau",email = "yasmin.venceslau@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-django (>=4.11.1,<5.0.0)"
]
